DRY visibility evaluator construction
