+++ kifi-backend/modules/search/app/com/keepit/search/engine/user/UserFromLibrariesScoreVectorSource.scala
@@ -46,9 +46,7 @@ class UserFromLibrariesScoreVectorSource(
     val pq = createScorerQueue(scorers, coreSize)
     if (pq.size <= 0) return // no scorer
 
-    val visibilityDocValues = reader.getNumericDocValues(LibraryFields.visibilityField)
-    val orgIdDocValues = reader.getNumericDocValues(LibraryFields.orgIdField)
-    val libraryVisibilityEvaluator = getLibraryVisibilityEvaluator(ownerIdDocValues, orgIdDocValues, visibilityDocValues)
+    val libraryVisibilityEvaluator = getLibraryVisibilityEvaluator(reader)
 
     val taggedScores: Array[Int] = pq.createScoreArray // tagged floats
